pyproject.toml
[tool.poetry]
name = "nome"
version = "0.1.0"
description = ""
authors = ["alessiogandelli <alessiogandelli99@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.0"
matplotlib = "^3.7.2"
python-dotenv = "^1.0.0"
jupyter = "^1.0.0"
python-telegram-bot = {extras = ["job-queue"], version = "^20.6"}
js2py = "^0.74"
ics = "^0.7.2"
asyncio = "^3.4.3"
schedule = "^1.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
